feat: view specific filters applied to a chart
