+++ packages/frontend/src/components/Explorer/FiltersCard/FiltersCard.styles.ts
@@ -1,3 +1,4 @@
+import { Colors, Tag } from '@blueprintjs/core';
 import styled from 'styled-components';
 
 export const CardHeader = styled.div`
@@ -18,4 +19,13 @@ export const Tooltip = styled.p`
 
 export const FilterValues = styled.span`
     font-weight: 700;
+`;
+
+export const TagContainer = styled(Tag)`
+    width: fit-content;
+    margin-right: 10px;
+
+    & span {
+        color: ${Colors.WHITE};
+    }
 `;
